Fixed a bug with market data values

+++ app/src/main/java/com/plcoding/cryptocurrencyappyt/data/remote/dto/CoinDetails/High24h.kt
@@ -54,7 +54,7 @@ data class High24h(
     val uah: Double,
     val usd: Double,
     val vef: Double,
-    val vnd: Int,
+    val vnd: Double,
     val xag: Double,
     val xau: Double,
     val xdr: Double,
